+++ release-notes.md
@@ -1,4 +1,15 @@
 # Tucana-20.16.0
+## Add kafka event notification for room entry updates ##
+- Issue Type: Task
+- Issue ID: FOUN-1370
+
+### Configuration ###
+    Replace all tokens (SERVICE_URL, ADMIN_CONSOLE_PARTNER_ADMIN_SECRET) from the template XML file below and remove ".template" from the file name:
+	/opt/kaltura/app/deployment/updates/scripts/xml/notifications/2024_03_03_add_kafka_room_entry_notifications.template.xml
+
+### Deployment scripts ###
+	php /opt/kaltura/app/deployment/updates/scripts/2024_03_03_deploy_kafka_room_events.php
+  
 # Add permissions for session get ##
 - Issue Type: Task
 - Issue ID: PLAT-24908